refactored functions to util module
