+++ ts_scripts/tsutils.py
@@ -2,7 +2,6 @@ import os
 import platform
 import time
 import requests
-import sys
 
 
 torchserve_command = {
@@ -10,14 +9,6 @@ torchserve_command = {
     "Darwin": "torchserve",
     "Linux": "torchserve"
 }
-
-
-def is_gpu_instance():
-    return True if os.system("nvidia-smi") == 0 else False
-
-
-def is_conda_env():
-    return True if os.system("conda") == 0 else False
 
 
 def start_torchserve(ncs=False, model_store="model_store", models="", config_file="", log_file="", wait_for=10):
@@ -85,10 +76,3 @@ def unregister_model(model_name, protocol="http", host="localhost", port="8081")
     url = f"{protocol}://{host}:{port}/models/{model_name}"
     response = requests.delete(url, verify=False)
     return response
-
-
-def check_python_version():
-    py_version = sys.version
-    if not py_version.startswith('3.'):
-        print("TorchServe supports Python 3.x only. Please upgrade")
-        exit(1)
